+++ R/connections.R
@@ -15,8 +15,8 @@
 #' @export
 #'
 #' @examples
-#' \dontrun{
-#' connection <- connect_biomart("ENSG")
+#' if(interactive()){
+#'   connection <- connect_biomart("ENSG")
 #' }
 connect_biomart <- function(prefix = c("ENSG", "GeneID")) {
   prefix <- match.arg(prefix)
@@ -30,7 +30,7 @@ connect_biomart <- function(prefix = c("ENSG", "GeneID")) {
 #' @aliases ConnectionBiomart
 #' @importClassesFrom biomaRt Mart
 #' @export
-.ConnectionBiomart <- setClass( # nolint
+.ConnectionBiomart <- setClass(
   "ConnectionBiomart",
   contains = "Mart",
   slots = c(prefix = "character")
@@ -43,9 +43,10 @@ connect_biomart <- function(prefix = c("ENSG", "GeneID")) {
 #' Helper function to query annotations from `biomaRt`, for cleaned up gene IDs of
 #' a specific ID variable and given [`biomaRt::Mart`][biomaRt::Mart-class].
 #'
-#' @param gene_ids (`character`)\cr gene IDs, e.g. `10329`.
+#' @param gene_ids (`character`)\cr gene IDs, e.g. `10329`, i.e. already
+#'   without the Entrez `GeneID` prefix, or `ENSG00000241644` for Ensembl gene ID.
 #' @param id_var (`string`)\cr corresponding gene ID variable name in BioMart,
-#'   e.g. `entrezgene_id`.
+#'   i.e. `entrezgene_id` or `ensembl_gene_id`.
 #' @param mart (`Mart`)\cr given [`biomaRt::Mart`][biomaRt::Mart-class] object.
 #'
 #' @return A data frame with columns:
@@ -53,17 +54,16 @@ connect_biomart <- function(prefix = c("ENSG", "GeneID")) {
 #'   - `hgnc_symbol`
 #'   - `entrezgene_description`
 #'   - `chromosome_name`
-#'   - `start_position`
-#'   - `end_position`
+#'   - `size`
 #'   - `refseq_mrna`
 #'   - `refseq_peptide`
 #'
 #' @export
 #'
 #' @examples
-#' \dontrun{
-#' mart <- biomaRt::useMart("ensembl", dataset = "hsapiens_gene_ensembl")
-#' h_get_annotation_biomart(c("11185", "10677"), id_var = "entrezgene_id", mart = mart)
+#' if(interactive()){
+#'   mart <- biomaRt::useMart("ensembl", dataset = "hsapiens_gene_ensembl")
+#'   h_get_annotation_biomart(c("11185", "10677"), id_var = "entrezgene_id", mart = mart)
 #' }
 h_get_annotation_biomart <- function(gene_ids,
                                      id_var,
@@ -78,14 +78,14 @@ h_get_annotation_biomart <- function(gene_ids,
       id_var,
       "hgnc_symbol",
       "entrezgene_description",
-      "chromosome_name",
-      "start_position",
-      "end_position"
+      "chromosome_name"
     ),
     filters = id_var,
     values = gene_ids,
     mart = mart
   )
+  gene_ids_values <- as.character(df_gene[[id_var]])
+  df_gene$size <- h_get_size_biomart(gene_ids_values, id_var = id_var, mart = mart)
   df_protein <- biomaRt::getBM(
     attributes = c(
       id_var,
@@ -99,7 +99,7 @@ h_get_annotation_biomart <- function(gene_ids,
   df <- merge(df_gene, df_protein, by = id_var, all = TRUE)
   df <- df[match(gene_ids, df[[id_var]]), ]
   rownames(df) <- gene_ids
-  df[, - which(colnames(df) == id_var)]
+  df[, -which(colnames(df) == id_var)]
 }
 
 # query ----
@@ -169,18 +169,158 @@ h_strip_prefix <- function(gene_ids, prefix) {
   )
 }
 
+#' Total Length of All Exons for Genes
+#'
+#' @description `r lifecycle::badge("experimental")`
+#'
+#' This helper function queries BioMart for lengths of genes by adding up all
+#' exon lengths after reducing overlaps.
+#'
+#' @inheritParams h_get_annotation_biomart
+#'
+#' @return Named integer vector indicating the gene lengths.
+#' @export
+#'
+#' @examples
+#' if(interactive()){
+#'   mart <- biomaRt::useMart("ensembl", dataset = "hsapiens_gene_ensembl")
+#'   h_get_size_biomart("11185", "entrezgene_id", mart)
+#'   h_get_size_biomart("ENSG00000215417", "ensembl_gene_id", mart)
+#'   h_get_size_biomart(c("11185", "10677"), "entrezgene_id", mart)
+#'   h_get_size_biomart(c("ENSG00000135407", "ENSG00000215417"), "ensembl_gene_id", mart)
+#' }
+h_get_size_biomart <- function(gene_ids,
+                               id_var,
+                               mart) {
+  assert_character(gene_ids)
+  assert_string(id_var)
+  assert_subset(id_var, choices = c("ensembl_gene_id", "entrezgene_id"))
+  assert_class(mart, "Mart")
+
+  # We can only query this starting with Ensembl Gene IDs.
+  attrs <- c(
+    "ensembl_gene_id",
+    "ensembl_exon_id",
+    "chromosome_name",
+    "exon_chrom_start",
+    "exon_chrom_end"
+  )
+  coords <- biomaRt::getBM(
+    filters = id_var,
+    attributes = attrs,
+    values = gene_ids,
+    mart = mart
+  )
+  ids <- unique(coords[, "ensembl_gene_id"])
+  granges_list <- sapply(ids, h_get_granges_by_id, coords = coords)
+  exons <- GenomicRanges::GRangesList(granges_list, compress = FALSE)
+  unique_exons <- GenomicRanges::reduce(exons)
+  unique_exon_sizes <- GenomicRanges::width(unique_exons)
+  total_exon_size <- sum(unique_exon_sizes)
+
+  if (id_var == "entrezgene_id") {
+    # Translate names back to Entrez gene IDs.
+    names(total_exon_size) <- h_ensembl_to_entrez_ids(
+      names(total_exon_size),
+      mart = mart
+    )
+  }
+
+  # Make sure to return the sizes in the correct order.
+  total_exon_size[gene_ids]
+}
+
+#' Translation of Ensembl to Entrez Gene IDs
+#'
+#' @description `r lifecycle::badge("experimental")`
+#'
+#' This helper function queries BioMart to translate Ensembl to Entrez Gene IDs.
+#'
+#' @param gene_ids (`character`)\cr Ensembl gene IDs.
+#' @inheritParams h_get_annotation_biomart
+#'
+#' @return Character vector of Entrez gene IDs.
+#' @export
+#'
+#' @examples
+#' if (interactive()) {
+#'   mart <- biomaRt::useMart("ensembl", dataset = "hsapiens_gene_ensembl")
+#'   h_ensembl_to_entrez_ids(c("ENSG00000135407", "ENSG00000241644"), mart)
+#' }
+h_ensembl_to_entrez_ids <- function(gene_ids,
+                                    mart) {
+  assert_character(gene_ids, pattern = "^ENSG")
+  assert_class(mart, "Mart")
+
+  translate_df <- biomaRt::getBM(
+    filters = "ensembl_gene_id",
+    attributes = c("ensembl_gene_id", "entrezgene_id"),
+    values = gene_ids,
+    mart = mart
+  )
+  result <- translate_df[match(gene_ids, translate_df$ensembl_gene_id), "entrezgene_id"]
+  as.character(result)
+}
+
+#' Conversion of BioMart Coordinates into `GRanges`
+#'
+#' @description `r lifecycle::badge("experimental")`
+#'
+#' This function extracts the chromosome number, the start position and the end position of transcripts
+#' in given `data.frame` with coordinates as returned by `biomaRt::getBM()` and converts
+#' them to a `GRanges` object.
+#'
+#' @param coords (`data.frame`)\cr as returned by `biomaRt::getBM()`, containing the columns
+#'   `ensembl_gene_id`, `chromosome_name`, `exon_chrom_start`, `exon_chrom_end`.
+#' @param id (`string`)\cr single Ensembl gene ID to convert the coordinates for.
+#'
+#' @return `GRange` objects for the respective single gene ID.
+#' @export
+#'
+#' @examples
+#' if(interactive()){
+#'   mart <- biomaRt::useMart("ensembl", dataset = "hsapiens_gene_ensembl")
+#'   attrs <- c("ensembl_gene_id",
+#'              "ensembl_exon_id",
+#'              "chromosome_name",
+#'              "exon_chrom_start",
+#'              "exon_chrom_end")
+#'   coords <- biomaRt::getBM(
+#'     filters = "entrezgene_id",
+#'     attributes = attrs,
+#'     values = c("11185", "10677"),
+#'     mart = mart
+#'   )
+#'   h_get_granges_by_id(coords, "ENSG00000135407")
+#' }
+h_get_granges_by_id <- function(coords, id) {
+  assert_data_frame(coords)
+  assert_names(
+    names(coords),
+    must.include = c("ensembl_gene_id", "chromosome_name", "exon_chrom_start", "exon_chrom_end")
+  )
+  assert_string(id)
+  has_id <- coords[, "ensembl_gene_id"] == id
+  assert_true(any(has_id))
+  id_exons <- coords[has_id, c("chromosome_name", "exon_chrom_start", "exon_chrom_end")]
+  GenomicRanges::GRanges(
+    id_exons$chromosome_name,
+    IRanges::IRanges(id_exons$exon_chrom_start, id_exons$exon_chrom_end)
+  )
+}
+
 # query-ConnectionBiomart ----
 
 #' @rdname query
 #'
 #' @export
 #' @examples
-#' \dontrun{
-#' object <- hermes_data
-#' connection <- connect_biomart(prefix(object))
-#' result <- query(genes(object), connection)
-#' head(result)
-#' head(annotation(object))
+#' if(interactive()){
+#'   object <- hermes_data
+#'   connection <- connect_biomart(prefix(object))
+#'   result <- query(genes(object), connection)
+#'   head(result)
+#'   head(annotation(object))
 #' }
 setMethod(
   f = "query",
@@ -206,7 +346,7 @@ setMethod(
         symbol = hgnc_symbol,
         desc = entrezgene_description,
         chromosome = as.character(chromosome_name),
-        size = end_position - start_position + 1L,
+        size = size,
         # Additional annotations.
         canonical_transcript = refseq_mrna,
         protein_transcript = refseq_peptide,